requirements.txt
channels-redis==4.0.0
coverage==7.1.0
dj-database-url==1.2.0
django-admin-interface==0.24.2
django-admin-sortable2==2.1.4
django-bootstrap-form==3.4
django-cors-headers==3.13.0
django-fernet-fields==0.6
django-hint==0.2.0
django-import-export==3.0.2
django-query-profiler==0.9
django-redis==5.2.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-storages==1.13.2
boto3==1.26.143
django-tinymce==3.5.0
djangorestframework-dataclasses==1.2.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-writable-nested==0.7.0
drf-yasg==1.21.4
faker==16.7.0
flake8==6.0.0
freezegun==1.2.2
grabzit==3.5.2
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==6.0.0
markuppy==1.14
more-itertools==9.0.0
numpy==1.24.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.0
phonenumbers==8.13.5
pip-chill==1.0.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pypdf2==3.0.1
pysocks==1.7.1
pytest-django==4.5.2
python-decouple==3.7
requests-pkcs12==1.15
service-identity==21.1.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
thefuzz==0.19.0
python-levenshtein==0.20.9
django-allauth==0.51.0
dj-rest-auth==3.0.0
websocket-client==1.5.1
pytest-asyncio==0.20.3
uvicorn[standard]==0.21.0
celery==5.3.4
django-celery-beat==2.5.0
flower==2.0.1
beautifulsoup4==4.12.2
deepl==1.16.1
lxml
pillow==10.1.0